srm: Fix regression in SRM startup
Motivation:

A change in 2.15 added cell lifecycle notifications. During startup
the cell is not available for communication with other cells. This
causes a regression in the SRM as it tries to restored requests during
startup and these requests may have expired. In that case the SRM tries
to clean up temporary upload directories in the name space, but this
blocks due to the cell communication not being available.

Modification:

Move the job restore until a phase after the cell has started.

Result:

Fixes a regression in the SRM that caused it to hang during startup.
Fixes #2354.

Target: trunk
Request: 2.15
Require-notes: yes
Require-book: no
Acked-by: Albert Rossi <arossi@fnal.gov>

Reviewed at https://rb.dcache.org/r/9198/