resilience: rename 'pnfs' command to 'file'
Motivation:

The designation 'pnfs' is vestigial and should not be
propagated by exposure through a user-facing interface
such as the admin commands.

In all the cases where this was used to name a command,
'file' is an appropriate alternative.

Modification:

pnfs=>file in commands and in the Java class names that
support them.

Result:

'Pnfs' is gone from the user interface.

Target: master
Depends-on: 9195
Acked-by: Gerd