resilience: rename 'pnfs' objects, fields and methods to 'file'
Motivation:

As with 9196, we should remove the vestigial uses of this designation
for operations on files.

Modification:

All references to 'PnfsOperation' or 'PnfsUpdate' have been replaced
with equivalents using 'File'.  This includes test code.  One enum
value was also changed from PNFSID to FILE.

There are lots of changes, but they are all renaming only.

Result:

No more 'pnfs{operation,update}' and derivatives in the code.

Target: master
Depends-on: 9196
Acked-by: Gerd