cells: Strip out source routing information when reverting a cell path
Motivation:

The need for source routing in cells has been greatly removed with the
simplification of the routing topology. Source routing now presents a
challenge as replies are source routed by default, forcing the reply
to go along the same route as the request. Thus despite multi pathing,
it may be impossible to receive a reply to a request if the core domain
contained in the route is no longer available.

Modification:

Collapse the source route when reverting it.  The PingMessage was extended to
preserve the source path. A couple of xyzzy use cases were converted to
PingMessage to make use of this information.

Result:

More robust multi pathing.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrchyan@desy.de>
Patch: https://rb.dcache.org/r/9203/

Reviewed at https://rb.dcache.org/r/9203/