cells: Controlled removal of downstream default routes in case of core domain shutdown
Motivation:

Shutdown in dCache used to not be particularly critical in the sense that when
central domains are shut down, dCache would not be functional anyway. Thus one
could safely kill connections to other domains without making things any worse.

Since the introduction of multi pathing, this is no longer the case. When
shutting down a core domain, the remaining domains may have alternate routes
to communicate. Thus how we decouple from the other domains matter.

A core domain cannot kill all tunnels exactly at the same time. During shutdown
there is a period in which the invariant that a core domain is connected to
all satellites is violated. Thus during this period one downstream could route
a message to the core domain that the core is unable to deliver.

Modification:

Adds a shutdown notification from cores to satellites. This notification is
sent to all attached satellites when a core domain shuts down, but before any
of the tunnels are killed. The core waits for a confirmation from all satellites
before severing any of the connections.

This logic is implemented in the CoreRoutingManager. To allow it to detect that
the domain is shutting down, the routing manager creates a canary cell. This
cell is a regular cell (that is, not a system cell) and is killed before any of
the tunnels are killed. The canary sends the notification to the satellites
when it is killed.

Upon receiving the notification, satelittes remove the corresponding default
route to the core.

For symmetry, the installation of a default route is moved from the tunnel into
the CoreRoutingManager. The delayed installation of all but the first default
route is refined such that if the last default route is removed, any delayed
default routes are installed right away.

Result:

More robust multi path messaging.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Albert Rossi <arossi@fnal.gov>

Reviewed at https://rb.dcache.org/r/9213/