resilience: eliminate the inaccessible file api and handler
Motivation:

The Inaccessible files API, implemented via a handler which writes
the pnfsids to a file named after the pool and which sends an alarm,
was originally conceived as an aid to administrators to warn them
when action would have to be taken out of band.

The adopted schema, however, has only limited usefulness, because
the list of pnfsids produced when a pool goes down or when no
readable/valid copies are discovered is not modified by the system
when a pool recovers.

To make this happen, a more complicated procedure would be necessary.
It would be possible, for instance, to implement the handler to
maintain in-memory mappings which are periodically written to file
and which can be reloaded on restart, removed entries from them
when pools which come back on line are scanned.  In other words,
duplicating the checkpointing functionality.

The latter was largely a concession to being able to run without
a database; to add yet another set of mechanisms would make this
service cumbersome, if not kludgy.

After due consideration, I have decided that a simple alarm
(by pool, with counter incrementing each time a new such file
is discovered) is sufficient.  The alarm contains further instructions
to run an admin command (to be added by a subsequent patch)
which will reveal which files are so orphaned on that pool.
Note that the alarm is sent during the "verify" phase of
the file operation, not during the pool scan itself, so
the alarm server will not be flooded with such messages
(maximum number of concurrent alarms would be equal to the
operation queue size, by default set to 200).

Modification:

The handler interface, implementation, and calls to it, along
with relevant unit tests and properties, have been removed from
the code base.  The simplified alarm has been inserted in the
appropriate location of the FileOperationHandler.

Result:

There is no effort to maintain (correctly and usefully)
a list of such files/pnfsids.

Target: master
Acked-by: Dmitry