webadmin: change PoolMonitorCollector to use SerializedPoolMonitor messages
Motivation:

The PoolManager now notifies on the pool monitor topic, so refreshes
of the PoolMonitor are uniformly pushed out to subscribers.  Hence
there should no longer be any need for ad hoc pulls of the monitor
via send and wait messages.

The webadmin Collector interface is designed to abstract such pull
requests, and the calling services rely on the call() interface
to gather information in this way.   It is neither trivial nor desirable to change
this paradigm within the webadmin service, but for the PoolMonitorCollector
it is nevertheless desirable not to make the PoolManager do extra work
which is unnecessary.

Modification:

The PoolMonitorCollector now merely recaches the refreshed PoolMonitor
object it gets from a receiver.  This receiver is injected into
the httpd webapp handler, and included in the JNDI configuration
which is made available to webadmin Spring injection; in this way
it is passed on to the PoolMonitorCollector.

Result:

No more ad hoc PoolMonitor request messages.

Changes are invisible to the user.

Target: master
Acked-by: Gerd