resilience: remove counts command
Motivation:

The counts command was intended to list file ids and
replica counts on a per-pool basis.  Its usefulness
is limited, however, for two reasons:  all it does
is wrap a query to the namespace (without any other
kind of filtering or verification), and thus can
be achieved equally well simply by executing such
a query; given the amount of time such a query
can take to run, especially on large namespaces,
tere is little advantage to running this
asynchronously (as one must) inside the resilience
service. (Note: while the same argument might
be made concerning the "inaccessible" command,
that command executes a more complicated,
three step query which also relies on PoolMonitor
information.)

Modification:

Eliminated the command, plus the two underlying
public methods in the namespace access interface
and implementation, since these are elsewhere
unused.  Note that some of the admin code
was reused to implement the new 'inaccessible'
command.

Result:

The counts command is gone, along with a good
bit of ad hoc sql inside the namespace class.

Target: master
Acked-by: Gerd