webdav: Changed the login strategy from Remote Login Strategy to CachingLoginStrategy
Currently all requests to the http end-point result in a call to gPlazma to verify the authentication credentials. This will be a considerable overhead with OpendID Connect because each verification will result in a call to each of the configured identity providers.

Ticket:
Acked-by: Gerd Behrmann <behrmann@gmail.com>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Target: trunk
Require-book: no
Require-notes: no
Patch: https://rb.dcache.org/r/9216/