webdav: Separate the authentication from checking of root/home path from AuthenticationHandler
Currently, the Authentication performs verification of authentication credentials and checks if the http request is within the root directory of user. This will be a problem with REST Api because there are nouns within the REST requests which may not map to a valid directory within the root director of user. In current scenario, the handler responds with a http redirect to a path which maps to a valid root path. This split will help the REST handler to run on a separate end-point but still use the AuthenticationHandler.

Ticket:
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Target: trunk
Require-book: no
Require-notes: no
Patch: https://rb.dcache.org/r/9218/