resilience: add command to find inaccessible files
Motiviation:

After the elimination of the inaccessible file handler api (see patch 9204),
we would still like to give the administrator some help in locating
files which are (currently) not accessible on the system.

Modification:

Added an admin command to run (asynchronously) a query against the namespace,
checking the locations of the given pnfsid to see if they are currently
readable, and printing to a file in resilience home the pnfsids for which
that collection of locations is empty.

Instructions to use this command are included in the alarm message sent
when no readable copies of a file are found.

Result:

Restored ability, albeit manual and upon request, to locate non-readable
files.

Target: master
Acked-by: Gerd