resilience: eliminate abstract commands and reintegrate annotation text
Motivation:

Earlier versions of the resilience system provided for an option
whereby the service could actually run embedded in PnfsManager.
This was undergoing testing, but was momentarily abandoned;
it is now clear that it will no longer be desirable or
necessary to provide for this option.

In connection with this, it made sense to prefix the admin
commands with something like 'res' when the commands were
exposed in the embedded version.  In order to achieve this,
it was necessary to push the code and the reusable text
for the hint and descriptions annotations into an
abstract class, and then simply to define the
class with its correct name annotation in a child concrete
class.  These were labeled EmbeddedResilienceCommands and
StandaloneResilienceCommands.

The former was eliminated for this release, but the underlying
abstract class and the single concrete class was retained,
just in case we might have wanted to revisit this code
(which will no longer be the case).

Hence this pattern has now become otiose.

Modification:

Make the base class final, incorporate the annotation text
directly into the definition, remove the static properties
and the child class.

Result:

Easier to read annotation code, and removal of unnecessary
inheritance structure.

Target: master
Acked-by: Gerd