webadmin: change pool queue plots to drill down from pool group data
Motivation:

Currently mover data for pools is plotted using a round-robin database
(rrd4j) and displayed on a single webadmin page.  For large systems,
this means displaying many plots in a single grid view.  While the filter
allows one to search for specific group names and such, it makes
more sense to organize this data hierarchically.  Moreover, an
administrator may be interested in breaking down pool queue activity
in terms of pool group; those sub-totals are not currently available.

Modification:

The plots are now organized in terms of three levels:  at the top
is the summary plot which gives an overview of the entire system.
Below that are the pool groups.  On each pool group plot there is
a link which allows one to display the individual pools in that
group.

Note:  I have eliminated the "all pools" view because of its
limited usefulness.

Implementation notes:

1.  all database files (both pools and groups) are stored at the top level
    of the plots directory.  Group plots are also at the top level.

2.  a group directory is created for each pool group, and in that directory
    are stored the pool plots.  If pools are shared there will be
    duplicate plot files (but only one database file).

3.  The panels and level settings (along with the filter, as previously)
    are stored in a session bean.  The refresh() method, which should
    be called before rendering, rebuilds the panels, so that on render,
    the provider merely fetches them.  For this reason, refresh()
    should continue to be called on each page reload (as before),
    so that up-to-date plots are seen (even if the filter or level
    has not changed).

A separate patch dependent on this one will add code to remove stale plots
and database files.

Result:

Easier to use display of plots.

Target: master
Require-notes: yes
Require-book: no
Acked-by: Gerd