webadmin: remove stale pool queue database files and plots
Motivation:

If pools or pool groups are eliminated from a running
dCache poolmanager configuration, their mover queue
plots should no longer be displayed.  Currently this
is not the case.

Modification:

Check for the existence of database files and image
files which do not correpond to pools or pool groups
transmitted with pool monitor updates.  These are
simply deleted (as this is ephemeral data anyway).

Result:

Up to date plots on the webadmin page.

Target: master
Depends-on: 9220
Require-notes: yes
Require-book: no
Acked-by: Gerd