poolmanager: fix double printout of poolmanager configuration
Motivation:

On a recent trunk build, 2.16@ce8e369, it was observed that issuing the "save" command in the PoolManager admin shell will duplicate entries in poolmanager.conf.

UniversalSpringCell calls printSetup for each bean which implements CellSetupProvider. In poolmanager.xml we create two beans: psu and PsuAccess. The second one is a decorator around psu. As a result, printSetup called twice.

Modification:

Eliminate the call to the delegate in the PsuAccess, as it is unnecessary. This had already been done in the beforeSetup and afterSetup methods; not doing so here was a simple oversight.

Result:

No duplicate entries, normal poolmanager.conf file on save.

Target: master
Acked-by: Gerd