srm: Split into frontend and backend services
Motivation:

The motivation for this patch goes all the way back to around 2008. At the
time BNL wanted to scale the SRM by separating the GSI authentication from
the SRM request processing, allowing the part with the authentication to
be horizontally scaled.

At the time we were using Tomcat, SRM was purely integrated with dCache,
and the design ideas revolved around using an external messaging system
to bind the frontends and backends. This never materialized.

The problem however persists. Despite improvements in the software and
vastly more powerful hardware, SRM authentication is expensive and forms
a hard limit on the request throughput of the SRM.

Modification:

Split the SRM service into two components: A frontend containing Jetty,
Axis, and the callout to gPlazma; and a backend containing the request
processing, including persistence in a database.

The frontend service is still called srm, while the backend is called
srmmanager. The srmmanager service is currently not scalable, but the
frontend can be scaled at will. An external load balancing mechanism
(round robind DNS, load balancing switch) has to be used to make the
frontends appear as one to the clients.

The frontend and backends are regular dCache services and communicate
over cells message passing.

Both the SRM and GridSite servlets are split in this manner.

Result:

Splits the srm service into two: A frontend service called srm, and a backend
service called the srmmanager. The frontend contains the Jetty webserver
and performs the authentication, while the backend contains all the
srm request processing and requires access to the srm database.

The frontend can be scalled at will, using a common backend.

Many configuration properties that used to belong to the srm service
are now obsolete and have been moved to the srmmanager service instead.

Note that it is the frontend that is advertised as a door in the info
service. For this reason both the frontends and the backend need to be
configured with a common name space root path.

It is no longer necessary to maintain a vomsdir configuration on neither
the frontend nor backend service. The srm frontend relies on gPlazma for
this functionality.

In contrast to before, the gridsite delegation service now requires that
the user is authorized through gPlazma.

Upon upgrade, as a minmum an srmmanager service needs to be added to the
layout file. It is recommended to place this before the srm service. If
any srm properties are set, these should be updated as necessary in case
these now belong to the srmmanager service.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>

Reviewed at https://rb.dcache.org/r/9243/