billing: modify insert to use single queue and thread pool executor service
Motivation:

Billing currently uses four consumer threads/queues, one for each of the
different messages which carry data to be inserted; each of these messages
maps, through the DAO layer, to a separate database table.  The separation
was an attempt to eliminate lock contention from simultaneous transactions
on a given table.

However, this scheme tends to limit throughput, even though a batched
commit (in DataNucleus-ese, "makePersistentAll") was being used.
The batching size can also be limiting, because
makePersistentAll is merely a loop wrapping individual writes of single
objects done serially. Multiple consumers on the queues seem to
be indicated.

Once we choose to go in that direction, however, the rationale for
maitaining separate queues and consumers disappears.  The additional
API layer composing a queue delegate also is an unnecessary abstraction.

Modification:

Insertion now takes place via a single queue with multiple consumer threads;
as before, the consumers drain the queue up to a max number of entries
and commit to the DataNucleus batch method. The queue and consumer logic
is all combined into a single abstract class, whose commit method is implemented
by the concrete class (using DataNucleus in this case).

Result:

Simpler design and better performance.

Target: master
Require-book: no
Require-notes: no
Acked-by: Dmitry
Acked-by: Gerd