pools: add CorruptFileTopic to excluded destinations on no route to cell
Motivation:

Resilience is the only subscriber to this topic.  If resilience is not enabled,
the failed delivery of the CorruptFileLocation message is logged as an error.

Modification:

Add this topic to the list of suppressed errors on no route to cell.

Result:

No more errors logged if resilience is not enabled.

Target: master
Acked-by: Gerd