srm: Fix regression in get request tokens and get request summary processing
Motivation:

The recent patch to load balance srm requests over multiple backends maps
requests by encoding the backend of a request in the request token. Two srm
opertions failed to be properly mapped as they operate on collections of
requests and thus failed to work with the simplistic mapping that had been
implemented. The bug affects srmGetRequestTokens and srmGetRequestSummary.

Modification:

srmGetRequestTokens is easily fixed by submitting the request to all backends
and aggregating the response.

srmGetRequestSummary is more complicated as the request contains a list of
request tokens that have to be grouped by backend, submitted to the relevant
backends, and aggregated on the frontend. This has to be done concurrently
for all backends. It is further complicated by the various combinations of
errors, such as some requests not being mappable to a backend, some backends
being offline, or that some backends may themselves report failures to
produce a request summary. The implementation is mostly asynchronous, using
collections and maps of listenable futures to capture the result for each
individual request token, and with code mapping the result of bulk requests
back to futures of individual request summaries.

Result:

Fixes a regression in processing srmGetRequestTokens and srmGetRequestSumary.

srmGetRequestTokens issued to a frontend srm service will fetch the list of
tokens from all available backends, no matter whether this frontend actually
submits to that backend. This is considered a feature rather than a bug as it
enables deployments of a cluster of load balanced frontends with a cluster of
backends where each frontend submits new scheduled requests to a single
(presumably local) backend only, yet can query the status of any request on
any of the backends. If this results in problems for a multi VO multi SRM
deployment, then srmGetRequestTokens has to be refined in the future.

Target: trunk
Require-notes: no
Require-book: no
Fixes: #2401
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9281/