resilience: add filter support for matching null pool values on file cancel and list commands
Motivation:

It is useful when listing and cancelling file operations to be
able to match not only determinate values for parent, source and
target pools, but also operations which have no value
(null) for those fields.

Modification:

Allow the file filter to handle empty string values.

Result:

The added functionality.

This patch also fixes an NPE bug in the origin match handling for
these fields.

Target: master
Acked-by: Gerd