resilience: fix handling of broken file where only one copy exists
When resilience receives notification of a broken (corrupt) file which is
under its control, it checks to see if this is the only copy of that file
(or if the file has not made it into the namespace).  If not, it removes
the copy and resets the operation to try (another source or target).

In the case of a single or non-existent file, it does nothing.  However,
it also needs to make sure that there is no running or waiting operation
connected to the pnfsid.  This (race) condition is not currently handled.

Modification:

In the case of a single/non-existent pnfsid, call cancel on the map and
force removal of the operation.

Result:

Avoiding a situation where futile retries keep occurring using the original
location.

Resilience has not yet been released, so no bug report is necessary.

Target: master
Acked-by:  Gerd