pnfsmanager: Fix precondition failure when querying the root path with PG backend
Motivation:

For postgresql and oracle, the chimera drivers produce an empty path when
querying the path of the root inode. Due to refactoring of FsPath, this is
now causing a precodition failure in FsPath.

Modification:

Fix the postgresql and oracle drivers to return "/" for the root inode. There
is no need for the overhead of actually querying the database in this case.

Result:

Fixes a regression in PnfsManager that would cause illegal state exceptions
when querying the path of the root inode.

Target: trunk
Require-notes: no
Require-book: no
Fixes: #2430
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9298/