pnfsmanager: change the timing of PnfsModifyCacheLocationMessage relays to follow reply to pools
Motivation:

Resilience subscribes to the PnfsAddCacheLocationMessage in order to get
notifications concerning new files written to pools.  This is like
the old Replica Manager.  However, because resilience operates entirely
in memory, response time to these notifications is much faster.

Under certain conditions, it is possible for resilience to begin a
copy request on a file before the pool has set the cache entry to
a state available for copying.  This is because PnfsManager relays
the PnfsAddCacheLocationMessages before the reply is sent to the pool.
In this case, the following error occurs:

16 May 2016 13:59:33 (dmsdca19-1.1) [RESILIENCE-aa314a15-e2e8-4e90-8a58-39b5de02c37b
Resilience PoolMigrationCopyReplica 000047E5291AF43F4DF0910B6A155F89E9BF]
P2P for 000047E5291AF43F4DF0910B6A155F89E9BF failed: Source pool failed (File is incomplete)

This provokes a retry inside of resilience.  While the retry usually
succeeds, it would be preferable to reduce the possibility of this
race occurring.

Note also that the PnfsManager receives PnfsAddCacheLocationMessages
from the pool in two cases:  a failure (WriteHandlerImpl.fail()),
and the execution of the 'pnfs register' command.  In these cases
the message is relayed before the message is actually queued for
processing.  But this case is not the crucial one.  Rather,
it is the new PnfsAddCacheLocationMessage(s) generated by
the PnfsSetFileAttributes message which seem to race the pool
state.

In either case, though, it would seem to make sense not to relay
these messages except from the pool consumer thread, and after
all processing, including sending the pool reply, has been
completed.

Modification:

All relays of this type of message are now done in one place:
in the postProcessLocationModificationMessage method, which calls
the relay method.   In the case of PnfsSetFileAttributeMessages,
the extraction of the locations is done here by first
setting the sent locations as a transient field and accessing
them afterwards.

Result:

Reduction of the number of errors of the type listed above.

Target: master
Request: 2.15
Require-notes: yes
Require-book: no
Acked-by:  Gerd