common: add parsing and formatting utilities for time duration
Motivation:

Being able to express a duration value as a formatted string
is useful in several places in dCache code.

Modification:

Add a generalized method to TimeUtils to return an object
encapsulating duration values accessible like a map.

Add a formatting class which allows for the expression of
duration in terms of days, hours, minutes, seconds and millis/nanos.

Result:

Standard methods for achieving this.

Target: master
Require-book: no
Require-notes: no
Acked-by: Dmitry