httpd: add VO-level information to Active Transfer representations
Motivation:

In dCache setups using VO authorization (e.g., VOMS/XACML), it is useful to be
able to track the primary group information (GID, FQAN) for the user.

Modification:

In order to display this information in the two web interface tables,
the Subject (already present elsewhere in the pertinent classes) needed
to be added to the IoDoorEntry.

Thereafter, the TransferObserver and corresponding webadmin classes had
to be adjusted to display the UID, GID and VOMS Group from the Subject.  The
'owner' field in the DoorInfo message is no longer used for this purpose.

Some effort has been made to render the code as DRY as possible.
JSON (along with html and text) output for the transfers
has been added to the TransferObserverV1 code.

Result:

Each transfer now reports UID, GID, and (when relevant) the FQAN group name.

NOTE:  The old active transfers page (which is still in place) displays
the replyHost; thus it defeats the purpose to hide this information from
the non-admin viewer in the webadmin page.  That column was restored to
the unauthorized view.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Gerd