webdav: add support for 'client' template in StringTemplate file
Motivation:

The new webdav client builds the user-visible page by querying a JSON
API directly.  This means that the information about the current
directory is not necessary and the current behaviour of the webdav door
where it queries PnfsManager for a directory listing only delays
delivering the content to the client.

Modification:

Add support for a 'client' template.  If the template-group file defines
a template called 'client' then this is used to build the response (the
entity) delivered to the client when a GET request is issued against a
directory.  This template will not have the directory contents.

If no such template exists, then the 'page' template is used to build
the HTML page from directory information from PnfsManager.

Result:

Faster loading times of the new JavaScript client, especially on busy
dCache instances for directories with large number of entries.

Target: master
Requires-notes: yes
Requires-book: no
Patch: https://rb.dcache.org/r/9319/
Acked-by: Gerd Behrmann