gplazma: Add PrefixRestriction and use it to validate root and upload paths
Motivation:

dCache user accounts can have a root path that restricts access for that user
to a particular part of the name space. Some doors also use this to alter the
name space root exposed through the door (the FTP door). The upload path is
a common exception to this path restriction as users should be able to upload
to this path even if it is outside the user's root directory.

The root path and upload path restrictions are implemented in every door. This
isn't DRY and comes with the risk that some code paths are forgotten. Further
some doors didn't even apply such checks, e.g. frontend and NFS.

Modification:

Introduce the PrefixRestriction to gPlazma for restricting access to paths that
have one of a given set of prefixes.

Update the gPlazma service to map the RootDirectory login attribute to a
PrefixRestriction. Adds the upload directory property to gPlazma and injects
this into the PrefixRestriction if necessary.

The RootDirectory login attribute is still present as some doors use this
not only for authorization but also to define the exposed name space root.

Stripped the duplicate authorization logic from doors.

The redirect to the user's home directory when accessing / and this being
outside the user's root directory has been refactored. This also eliminates
the need for a special Jetty handler to implements this.

Result:

The following properties are obsolete: ftp.authz.upload-directory,
webdav.authz.upload-directory, xrootd.authz.upload-directory. The following
property has been introduced: gplazma.authz.upload-directory.

Adds root path and upload path checks to services that did not have them before
(dcap, nfs, frontend). Request backport to 2.16 as the patch sort of fixes
an authorization issue.

Target: trunk
Request: 2.16
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9347/

(cherry picked from commit a09a461c98b9eebbeaa3b5e3046a623623bf98ff)