billing: use in-memory buffer for hourly aggregate data
Motivation:

Hourly aggregate data (used to create the 24-hour billing plots)
is currently served from database views based on the billinginfo
table.   It has been observed, however, that when throughput is
high over a 24-hour period (in the millions of inserts per hour),
there is a slowdown in response from the queries accessing these
views, sometimes provoking a timeout on the request.

Two things can be done to remedy this situation.  First,
the caller (in this case, the httpd plotting client) could be
rewritten to use delayed callbacks.  Second, latency can be
reduced by holding the hourly data in memory.  The latter makes
sense inasmuch as the window for this data is small.

Modification:

In this patch we have implemented a circular buffer of 24 hour-long
data bins (using a linked deque).  The buffer is populated
on (re)start from the original database views which we have
left in place.  The billing service no longer blocks on database
IO when serving up the hourly aggregate data (though it still
does on the other aggregate tables, but latency there is
negligible); on the initial request, the histogram objects
may thus carry empty or partial data, but the plots will eventually
be fully populated during subsequent refreshing.

We leave the httpd client changes for another patch. While
hourly data was taking considerably longer to serve
than the daily, weekly or yearly data, it is still a good
idea to change the client to be asynchronous.

Result:

Lower latency in requesting histogram data from billing.

Target: master
Request: 2.16
Request: 2.15
Request: 2.14
Request: 2.13
Require-book: no
Require-notes: yes
Acked-by: Gerd Behrmann behrmann@ndgf.org