pool: Add open flush mode
Motivation:

There are several use cases in which the batching of flush requests in a pool
is a problem. Typically some external system employs batching already and thus
the batching in dCache may prevent the external system from "seeing" enough
files to trigger a flush.

The core of the problem is that the batching logic as implemented in dCache
waits for a sufficient number of files to accumulate and then flushes that
batch without flushing any other files in the same storage class until that
entire batch has been flushed.

Modification:

Adds an open flush option to storage class flush queues. When in open flush
mode, additional files that are added to the pool after the flush of the
current batch was initiated will be flushed right away.

In case several storage class queues are defined, the flushing queue will be
put into a draining mode when the flush trigger condition of one of the other
queues is satisfied. In draining mode newly added files will not be flushed.
This arvoids starvation problems.

Result:

Added an `-open` option to the `queue define class` command in the `pool`
service.  This option alters the batching logic such that newly added files are
added to an existing flush operation. This is particularly useful in case an
external system is doing its own batching.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9379/