REST-api: fix permission denied.
Motivation:

Attempts to query a resource that the user is not allowed result in a
405 method not allowed response.  This is plain wrong and causes dCache
View not to handle the situation correctly.

Modification:

The response if the user attempts to list or stat content that they are
not allowed to see is updated.  If the user is logged in then a 403
Forbidden is returned; if the user is not logged in then a 401
Unauthorized is returned.

Result:

More correct responses; hopefully dCacheView working better.

Target: master
Request: 2.16
Requires-notes: true
Requires-book: false
Patch: https://rb.dcache.org/r/9388/
Acked-by: Gerd Behrmann