resilience: fail when source retry is requested with only one possible source
Motivation:

When a source is selected by resilience for copying, a check
is also made to see that it is not taken from among the
locations that have already been (unsuccessfully) tried.

The current code, however, does not do this in the case where
only one location exists.

Modification:

Always check for the tried locations.  In the case of a single
location, throw an exception.

Result:

No more bug.

Target: master
Request: 2.16
Require-notes: yes
Require-book: no
Acked-by: Gerd